first dev for flask app

+++ plugins/imaging_methods/HadamardSplit/PatternsCreation.py
@@ -58,7 +58,7 @@ class CreationPatterns:
         """
         hadamard_matrix = hadamard(
             self.dim**2, dtype=np.int8
-        )  # initialized an dim by dim Hadamard matrix
+        ).astype(np.int16)  # initialized an dim by dim Hadamard matrix
         self.sequence = []
         for col in range(self.dim**2):
             self.sequence.append(
